requirements.txt
anyio==3.7.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
urllib3==2.0.3
xmltodict==0.13.0
